exec: split sync+async code to separate files.
